Update getObjectValueByDotSeparatedKey; rename logError to serverLog and adjust formatting
In getObjectValueByDotSeparatedKey I improved the type safety. You now
need to pass a generic leaf node
(https://en.wikipedia.org/wiki/Tree_(data_structure)) type parameter to
it. I also clarified that this is meant to be used only on the server —
it will now log an error if called from the browser.

In internal projects we’ll probably wrap this in most cases (see
getStringByDotSeparatedKey in the project we’re testing this in).